# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Republic API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /websites:
    get:
      operationId: listWebsites
      tags: [websites]
      summary: List websites
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: isFree
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Website"
  /websites/{id}:
    get:
      operationId: getWebsite
      tags: [websites]
      summary: Get a single website
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
        "404":
          description: Not found

  # ── User ─────────────────────────────────────────────────────────────────────
...
  /support/tickets:
    get:
      operationId: listSupportTickets
      tags: [support]
      summary: List user support tickets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createSupportTicket
      tags: [support]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"

  # ── Notifications ─────────────────────────────────────────────────────────────
...
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      summary: Admin list all categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    patch:
      operationId: adminUpdateCategory
      tags: [admin]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/reorder:
    post:
      operationId: adminReorderCategories
      tags: [admin]
      summary: Reorder categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryReorderInput"
      responses:
        "200":
          description: OK

  # ── Admin: Websites ───────────────────────────────────────────────────────────
  /admin/websites:
    get:
      operationId: adminListWebsites
      tags: [admin]
      summary: Admin list all websites
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Website"
    post:
      operationId: adminCreateWebsite
      tags: [admin]
      summary: Create a website
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
  /admin/websites/{id}:
    patch:
      operationId: adminUpdateWebsite
      tags: [admin]
      summary: Update a website
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"

  # ── Admin: Users ──────────────────────────────────────────────────────────────
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin list all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{userId}/membership:
    patch:
      operationId: adminUpdateUserMembership
      tags: [admin]
      summary: Update user membership
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"

  # ── Admin: Support ────────────────────────────────────────────────────────────
  /admin/support/tickets:
    get:
      operationId: adminListSupportTickets
      tags: [admin]
      summary: Admin list all support tickets
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
  /admin/support/tickets/{id}:
    patch:
      operationId: adminUpdateSupportTicket
      tags: [admin]
      summary: Update support ticket (reply, close)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketAdminUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"

  # ── Admin: Stats ──────────────────────────────────────────────────────────────
...
  /admin/webview-settings:
    get:
      operationId: adminGetWebviewSettings
      tags: [admin]
      summary: Get global WebView settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebviewSettings"
    patch:
      operationId: adminUpdateWebviewSettings
      tags: [admin]
      summary: Update global WebView settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebviewSettingsUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebviewSettings"

  # ── Admin: Audit Logs ─────────────────────────────────────────────────────────
  /admin/audit-logs:
    get:
      operationId: adminListAuditLogs
      tags: [admin]
      summary: List audit logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogList"

  # ── Admin: Notifications ──────────────────────────────────────────────────────
  /admin/notifications:
    post:
      operationId: adminSendNotification
      tags: [admin]
      summary: Send notification to user(s)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Created
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, isActive, sortOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        sortOrder:
          type: integer
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
    CategoryReorderInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer

    # ── Website ──────────────────────────────────────────────────────────────────
    Website:
      type: object
      required: [id, name, url, displayDomain, isFree, isActive, categoryId, tabOrder, cardOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        displayDomain:
          type: string
        iconUrl:
          type: ["string", "null"]
        isFree:
          type: boolean
        isActive:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        tabOrder:
          type: integer
        cardOrder:
          type: integer
        canBeTab:
          type: boolean
        canPreload:
          type: boolean
        remembersUrl:
          type: boolean
        customUserAgent:
          type: ["string", "null"]
        injectedCss:
          type: ["string", "null"]
        injectedJs:
          type: ["string", "null"]
        appBannerHideSelectors:
          type: ["string", "null"]
        popupMitigationEnabled:
          type: boolean
        notes:
          type: ["string", "null"]
    WebsiteInput:
      type: object
      required: [name, url, displayDomain]
      properties:
        name:
          type: string
        url:
          type: string
        displayDomain:
          type: string
        iconUrl:
          type: ["string", "null"]
        isFree:
          type: boolean
        isActive:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        tabOrder:
          type: integer
        cardOrder:
          type: integer
        canBeTab:
          type: boolean
        canPreload:
          type: boolean
        remembersUrl:
          type: boolean
        customUserAgent:
          type: ["string", "null"]
        injectedCss:
          type: ["string", "null"]
        injectedJs:
          type: ["string", "null"]
        appBannerHideSelectors:
          type: ["string", "null"]
        popupMitigationEnabled:
          type: boolean
        notes:
          type: ["string", "null"]
    WebsiteUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        displayDomain:
          type: string
        iconUrl:
          type: ["string", "null"]
        isFree:
          type: boolean
        isActive:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        tabOrder:
          type: integer
        cardOrder:
          type: integer
        canBeTab:
          type: boolean
        canPreload:
          type: boolean
        remembersUrl:
          type: boolean
        customUserAgent:
          type: ["string", "null"]
        injectedCss:
          type: ["string", "null"]
        injectedJs:
          type: ["string", "null"]
        appBannerHideSelectors:
          type: ["string", "null"]
        popupMitigationEnabled:
          type: boolean
        notes:
          type: ["string", "null"]

    # ── User ─────────────────────────────────────────────────────────────────────
...
    Membership:
      type: object
      required: [userId, plan, status]
      properties:
        userId:
          type: string
        plan:
          type: string
          enum: [free, monthly, annual]
        status:
          type: string
          enum: [active, canceled, past_due, none]
        stripeCustomerId:
          type: ["string", "null"]
        stripeSubscriptionId:
          type: ["string", "null"]
        currentPeriodEnd:
          type: ["string", "null"]
    MembershipUpdate:
      type: object
      properties:
        plan:
          type: string
        status:
          type: string
...
    SupportTicket:
      type: object
      required: [id, type, subject, message, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        userEmail:
          type: ["string", "null"]
        type:
          type: string
          enum: [support, bug, feature]
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        adminReply:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    SupportTicketInput:
      type: object
      required: [type, subject, message]
      properties:
        type:
          type: string
          enum: [support, bug, feature]
        subject:
          type: string
        message:
          type: string
    SupportTicketAdminUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        adminReply:
          type: ["string", "null"]

    # ── Notifications ─────────────────────────────────────────────────────────────
...
    NotificationInput:
      type: object
      required: [title, message]
      properties:
        userId:
          type: ["string", "null"]
        title:
          type: string
        message:
          type: string

    # ── Admin ─────────────────────────────────────────────────────────────────────
    AdminUser:
      type: object
      required: [id, email, displayName, isPro, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        isPro:
          type: boolean
        membershipPlan:
          type: string
        membershipStatus:
          type: string
        createdAt:
          type: string
    AdminUserList:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
...
    WebviewSettings:
      type: object
      required: [maxPreloadedWebviews, preloadEnabled, sessionMemoryEnabled, homepageLimit, freeWebsiteLimit, proWebsiteLimit]
      properties:
        maxPreloadedWebviews:
          type: integer
        preloadEnabled:
          type: boolean
        sessionMemoryEnabled:
          type: boolean
        homepageLimit:
          type: integer
        freeWebsiteLimit:
          type: integer
        proWebsiteLimit:
          type: integer
    WebviewSettingsUpdate:
      type: object
      properties:
        maxPreloadedWebviews:
          type: integer
        preloadEnabled:
          type: boolean
        sessionMemoryEnabled:
          type: boolean
        homepageLimit:
          type: integer
        freeWebsiteLimit:
          type: integer
        proWebsiteLimit:
          type: integer
    AuditLog:
      type: object
      required: [id, action, createdAt]
      properties:
        id:
          type: integer
        adminId:
          type: string
        action:
          type: string
        details:
          type: ["string", "null"]
        createdAt:
          type: string
    AuditLogList:
      type: object
      required: [logs, total]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer