# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Republic API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditLog {
  id: number;
  adminId?: string;
  action: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}