# lib/api-zod/src/generated/types/auditLogList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Republic API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLog } from './auditLog';

export interface AuditLogList {
  logs: AuditLog[];
  total: number;
}