# lib/api-zod/src/generated/types/websiteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Republic API
 * OpenAPI spec version: 0.1.0
 */

export interface WebsiteInput {
  name: string;
  url: string;
  displayDomain: string;
  /** @nullable */
  iconUrl?: string | null;
  isFree?: boolean;
  isActive?: boolean;
  /** @nullable */
  categoryId?: number | null;
  tabOrder?: number;
  cardOrder?: number;
  canBeTab?: boolean;
  canPreload?: boolean;
  remembersUrl?: boolean;
  /** @nullable */
  customUserAgent?: string | null;
  /** @nullable */
  injectedCss?: string | null;
  /** @nullable */
  injectedJs?: string | null;
  /** @nullable */
  appBannerHideSelectors?: string | null;
  popupMitigationEnabled?: boolean;
  /** @nullable */
  notes?: string | null;
}