# lib/api-zod/src/generated/types/webviewSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Republic API
 * OpenAPI spec version: 0.1.0
 */

export interface WebviewSettings {
  maxPreloadedWebviews: number;
  preloadEnabled: boolean;
  sessionMemoryEnabled: boolean;
  homepageLimit: number;
  freeWebsiteLimit: number;
  proWebsiteLimit: number;
}