# lib/api-zod/src/generated/types/webviewSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Republic API
 * OpenAPI spec version: 0.1.0
 */

export interface WebviewSettingsUpdate {
  maxPreloadedWebviews?: number;
  preloadEnabled?: boolean;
  sessionMemoryEnabled?: boolean;
  homepageLimit?: number;
  freeWebsiteLimit?: number;
  proWebsiteLimit?: number;
}